+++ release_notes.md
@@ -20,6 +20,19 @@
 - Improve how certificates are evaluated for the authentication types `SAMLAssertion` and `OAUTH2_SAML_BEARER_ASSERTION`.
   The certificates are no longer unnecessarily parsed when obtaining the destination.
 - The `spring-boot3` archetype no longer contains the `integration-tests` module. The contained tests are now part of the test suite of the `application` module.
+- OpenAPI clients can now be generated with specifications with `oneOf` and `anyOf` keywords by enabling their processing by using `<enableOneofAnyofGeneration>` in `openapi-generator-maven-plugin`(This option is turned off by default):
+  ```xml
+  <plugin>
+      <groupId>com.sap.cloud.sdk.datamodel</groupId>
+      <artifactId>openapi-generator-maven-plugin</artifactId>
+      ...   
+      <configuration>
+         ...
+        <enableOneofAnyofGeneration>true</enableOneofAnyofGeneration>
+      </configuration>
+  </plugin>
+  ```
+  But,the generated client may not be feature complete and work as expected for all cases involving `anyOf/oneOf`.
 - Dependency Updates:
   - Update `org.apache.olingo` from `4.10.0` to `5.0.0`
   - Update `io.github.resilience4j` from `1.7.1` to `2.2.0`